requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
blinker==1.6.3
cachetools==5.3.2
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.6
django-allauth==0.58.1
django-cors-headers==4.3.1
django-filter==23.3
django-jazzmin==2.6.0
django-rest-passwordreset==1.3.0
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
Flask==3.0.0
gevent==23.9.1
google-auth==2.23.4
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
msgpack==1.0.7
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
reportlab==4.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
users==1.0.dev0
vine==5.1.0
wcwidth==0.2.11
websocket==0.2.1
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
